authenticate /nodes with jwt
